+++ spec/factories/ios_locations.rb
@@ -1,6 +1,6 @@
 FactoryGirl.define do
-  factory :ios_location do
-    name 'berlin'
+  factory :ios_location, aliases: [:location] do
+    name 'Acapulco'
     creator
   end
 end