+++ spec/models/user_spec.rb
@@ -14,7 +14,6 @@ RSpec.describe User, type: :model do
   it { is_expected.to respond_to(:deleted_at) }
   it { is_expected.to respond_to(:std_slots) }
   it { is_expected.to respond_to(:re_slots) }
-  it { is_expected.to respond_to(:location_name) }
   it { is_expected.to respond_to(:device_token) }
   it { is_expected.to have_many(:images) }
   it { is_expected.to have_many(:created_slots).inverse_of(:creator) }
@@ -27,6 +26,7 @@ RSpec.describe User, type: :model do
   it { is_expected.to have_many(:group_slots).through(:groups) }
   it { is_expected.to have_many(:initiated_friendships).inverse_of(:user) }
   it { is_expected.to have_many(:received_friendships).inverse_of(:friend) }
+  it { is_expected.to belong_to(:location) }
 
   it { is_expected.to be_valid }