+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 #Changelog
 
+## v0.4.0 on t2s2-silvio-dev
+- accept empty endDate for slots. It will internally be set to the end of the
+start day if start before midday and to middle of next day if start_date after
+midday, but end_date will not be returned via JSON
+- user has now an ios_location as home address (tough it's only called 'location'
+- json now mostly has nested objects instead of only the IDs
+user: locationId
+slot: creatorId, groupId and locationId
+group: groupId => id, ownerId
+- update ERDs
+
+
 ## v0.3.0 on t2s2-dev
 - use Puma as Webserver on Heroku instead of Unicorn